requirements (1).txt
pandas
sklearn
streamlit 
numpy
request
matplotlib
